Change directory structure for cli
